+++ CHANGELOG.md
@@ -11,6 +11,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Fix potential bugs related to system settings
 - Fix NoSuchFileException when output parent directories are nonexistent [#342](https://github.com/nbbrd/java-console-properties/issues/342)
+- Fix file ordering in MultiFileInput [#343](https://github.com/nbbrd/java-console-properties/issues/343)
 
 ## [1.4.0] - 2022-10-19